move the DenomOwners to uibc module
