+++ proto/umee/uibc/v1/query.proto
@@ -5,6 +5,8 @@ import "cosmos_proto/cosmos.proto";
 import "gogoproto/gogo.proto";
 import "google/api/annotations.proto";
 import "google/protobuf/timestamp.proto";
+import "cosmos/bank/v1beta1/query.proto";
+
 import "umee/uibc/v1/quota.proto";
 import "umee/uibc/v1/uibc.proto";
 
@@ -39,6 +41,12 @@ service Query {
   // QuotaExpires returns when current ibc quota will end.
   rpc QuotaExpires(QueryQuotaExpires) returns (QueryQuotaExpiresResponse) {
     option (google.api.http).get = "/umee/uibc/v1/quota_expires";
+  }
+
+  // DenomOwners queries for all account addresses that own a particular token
+  // denomination.
+  rpc DenomOwners(cosmos.bank.v1beta1.QueryDenomOwnersRequest) returns (cosmos.bank.v1beta1.QueryDenomOwnersResponse){
+    option (google.api.http).get = "/umee/uibc/v1/denom_owners";
   }
 }
