+++ x/ugov/keeper/intest/keeper.go
@@ -7,7 +7,6 @@ import (
 	cdctypes "github.com/cosmos/cosmos-sdk/codec/types"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	bkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
 
 	"github.com/umee-network/umee/v6/tests/tsdk"
 	"github.com/umee-network/umee/v6/x/ugov"
@@ -21,7 +20,7 @@ func MkKeeper(t *testing.T) (*sdk.Context, ugov.Keeper) {
 	ugov.RegisterInterfaces(ir)
 	cdc := codec.NewProtoCodec(ir)
 	storeKey := storetypes.NewMemoryStoreKey(ugov.StoreKey)
-	kb := keeper.NewBuilder(cdc, storeKey, bkeeper.BaseKeeper{}, true)
+	kb := keeper.NewBuilder(cdc, storeKey)
 	ctx, _ := tsdk.NewCtxOneStore(t, storeKey)
 	return &ctx, kb.Keeper(&ctx)
 }
