+++ x/uibc/expected_keepers.go
@@ -1,6 +1,8 @@
 package uibc
 
 import (
+	context "context"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/bank/types"
 
@@ -12,6 +14,10 @@ type BankKeeper interface {
 	GetDenomMetaData(ctx sdk.Context, denom string) (types.Metadata, bool)
 	SetDenomMetaData(ctx sdk.Context, denomMetaData types.Metadata)
 	IterateAllDenomMetaData(ctx sdk.Context, cb func(types.Metadata) bool)
+	DenomOwners(
+		goCtx context.Context,
+		req *types.QueryDenomOwnersRequest,
+	) (*types.QueryDenomOwnersResponse, error)
 }
 
 type Leverage interface {
